+++ R/updates_and_reports.R
@@ -18,16 +18,17 @@ nflverse_packages <- function(include_self = FALSE) {
   sort(names)
 }
 
-#' @inherit nflreadr::nflverse_sitrep
+#' Get a Situation Report on System, nflverse Package Versions and Dependencies
+#'
+#' @description This function gives a quick overview of the versions of R and
+#'   the operating system as well as the versions of nflverse packages, options,
+#'   and their dependencies. It's primarily designed to help you get a quick
+#'   idea of what's going on when you're helping someone else debug a problem.
+#' @details See [`nflreadr::nflverse_sitrep`] for details.
 #' @export
-#' @returns Returns `NULL` invisibly. Called for side effects.
-#' @examples
-#' \donttest{
-#' try(
-#' nflverse_sitrep()
-#' )
-#' }
-nflverse_sitrep <- nflreadr::nflverse_sitrep
+#' @name nflverse_sitrep
+#' @importFrom nflreadr nflverse_sitrep
+NULL
 
 #' Update nflverse Packages
 #'
@@ -54,10 +55,17 @@ nflverse_sitrep <- nflreadr::nflverse_sitrep
 nflverse_update <- function(recursive = FALSE,
                             repos = getOption("repos"),
                             devel = FALSE){
-  devel_repo <- "https://nflverse.r-universe.dev/"
-  if(isTRUE(devel)) repos["devel"] <- devel_repo
+  opts <- NULL
+  if(isTRUE(devel)){
+    repos["nflverse"] <- "https://nflverse.r-universe.dev/"
+    opts <-
+      'options(repos = c(
+        nflverse = "https://nflverse.r-universe.dev",
+        getOption("repos")
+      ))'
+  }
   available <- utils::available.packages(repos = repos)
-  packages <- nflverse_packages(include_self = FALSE)
+  packages <- nflverse_packages(include_self = TRUE)
 
   if(isTRUE(recursive)){
     deps <- tools::package_dependencies(packages, db = available) %>%
@@ -104,13 +112,12 @@ nflverse_update <- function(recursive = FALSE,
                   " -> ",
                   format(behind$cran), ")"
                   )
-  cli::cli_alert_info("Start a clean R session then run:")
   pkg_str <- paste0(deparse(behind$package), collapse = "\n")
-  if(isTRUE(devel)){
-    cli::cli_text("{.code install.packages({pkg_str}, repos = \"{devel_repo}\")}")
-  } else {
-    cli::cli_text("{.code install.packages({pkg_str})}")
-  }
+  out_string <- paste0("install.packages(", pkg_str, ')')
+
+  cli::cli_rule(left = "{.emph Start a clean R session then run}")
+  cli::cli_code(opts, out_string)
+  cli::cli_rule()
 
   invisible()
 }